Django mini projects/django_elasticsearch/requirements.txt
django
djangorestframework
django-rest-auth
django-import-export
uritemplate
swagger-ui
psycopg2-binary
django_opentracing
jaeger-client